PartitionDxe/Mbr : MdeModulePkg : added break to handle invalid LBA0 in MBR
modified:   MdeModulePkg/Universal/Disk/PartitionDxe/Mbr.c
Read Disk does a modification of ExtMbrStartingLba with the code MultU64x32 (ExtMbrStartingLba, BlockSize)
Error detection to see if ExtMbrStartingLBA has a value of 0. This is
invalid as LBA 0 = MBR.
After modification, the next time ExtMbrStartingLba is in this function
if ExtMbrStartingLba is set to 0 in the MBR it never passes the while/do
evaluation
It is multiplied by 0 by read disk , set to 0 by an invalid MBR and goes
back to evaluation
This condition will also cause Ws19 and WS22 to hang, however Microsoft
has developed a hotfix patch that will be released in 2022

CC: Liming Gao <gaoliming@byosoft.com.cn> [lgao4]
CC: Jian J Wang <jian.j.wang@intel.com> [jwang36]
CC: Hao A Wu <hao.a.wu@intel.com> [hwu25]
CC: Ray Ni <ray.ni@intel.com> [niruiyu]
CC: Zhichao Gao <zhichao.gao@intel.com> [ZhichaoGao]

Signed-off-by: Craig Edwards <craig.edwards@dell.com.

Developer's Certificate of Origin 1.1

By making a contribution to this project, I certify that:

(a) The contribution was created in whole or in part by me and I
    have the right to submit it under the open source license
    indicated in the file; or

(b) The contribution is based upon previous work that, to the best
    of my knowledge, is covered under an appropriate open source
    license and I have the right under that license to submit that
    work with modifications, whether created in whole or in part
    by me, under the same open source license (unless I am
    permitted to submit under a different license), as indicated
    in the file; or

(c) The contribution was provided directly to me by some other
    person who certified (a), (b) or (c) and I have not modified
    it.

(d) I understand and agree that this project and the contribution
    are public and that a record of the contribution (including all
    personal information I submit with it, including my sign-off) is
    maintained indefinitely and may be redistributed consistent with
    this project or the open source license(s) involved.

Copyright © 2021 Dell Inc. or its subsidiaries.  All Rights Reserved.

Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:

1. Redistributions of source code must retain the above copyright notice, this list of conditions and the following disclaimer.

2. Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the following disclaimer in the documentation and/or other materials provided with the distribution.

3. Neither the name of the copyright holder nor the names of its contributors may be used to endorse or promote products derived from this software without specific prior written permission.

THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Date:      Wed Dec 8 13:48:32 2021 -0600

 On branch graceful_handle_mbr_hang
 Changes to be committed:
	modified:   MdeModulePkg/Universal/Disk/PartitionDxe/Mbr.c